pyproject.toml
[project]
name = "llm-council"
version = "0.1.0"
description = "Your LLM Council"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "ollama>=0.3.0",
    "groq>=0.9.0",
    "google-generativeai>=0.7.0",
    "chromadb>=0.5.0",
    "DrissionPage>=4.0.0",
    # Hybrid AI Council additions
    "langgraph>=0.2.0",
    "ddgs>=7.0.0",
    "markitdown>=0.0.1",
    "python-multipart>=0.0.9",
    "psutil>=5.9.0",
]
